+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # `wasm-bindgen` Change Log
 --------------------------------------------------------------------------------
 
+## Unreleased
+
+### Changed
+
+* Add clear error message to communicate new feature resolver version requirements.
+  [#4312](https://github.com/rustwasm/wasm-bindgen/pull/4312)
+
+--------------------------------------------------------------------------------
+
 ## [0.2.97](https://github.com/rustwasm/wasm-bindgen/compare/0.2.96...0.2.97)
 
 Released 2024-11-30